Move IrcServer.connectAs to VirtualIrcUser.connectAs. Add names.js
The new file names.js is intended to contain all the IRC nick/uname/realname
logic in terms of mapping to LONGNAM~1 and stripping bad chars. Currently it
is just a stub file which does the bit of nick pre-processing which
connectAs was previously doing.
