+++ lib/irclib/client.js
@@ -1,7 +1,11 @@
 "use strict";
 
 var q = require("q");
+var ident = require("./ident");
+var actions = require("../models/actions");
+var clientConnection = require("./client-connection");
 var pool = require("./server-pool");
+var names = require("./names");
 var IrcRoom = require("../models/rooms").IrcRoom;
 var log = require("../logging").get("irc-client");
 
@@ -23,7 +27,7 @@ function VirtualIrcUser(ircUser, userId, isBot) {
 
 VirtualIrcUser.prototype.connect = function(hooks) {
     var that = this;
-    return this.server.connectAs({
+    return this.connectAs({
         nick: this.nick,
         username: this.userId
     }, hooks).then(function(connInst) {
@@ -64,6 +68,188 @@ VirtualIrcUser.prototype.connect = function(hooks) {
             }
         });
     });
+};
+
+/**
+ * @param {Object} connectionOpts
+ * @param {Object} callbacks
+ * @return {ConnectionInstance} A new connected connection instance.
+ */
+VirtualIrcUser.prototype.connectAs = function(connectionOpts, callbacks) {
+    var channels = connectionOpts.channels;
+    var server = this.server;
+    var nick = connectionOpts.nick;
+    var clientIsBot = nick === server.nick;
+
+    var nameInfo = names.getIrcNames(
+        this.server, this.nick, connectionOpts.username, this.userId
+    );
+
+
+    var password = clientIsBot ? server.nickPass : undefined;
+
+    // strip do not track channels
+    if (channels) {
+        for (var i = 0; i < channels.length; i++) {
+            if (server.doNotTrackChannels.indexOf(channels[i]) !== -1) {
+                channels.splice(i, 1);
+                i--;
+            }
+        }
+    }
+
+    log.info("Connecting to IRC server %s as %s (user=%s)- Joining channels %s",
+        server.domain, nameInfo.nick, nameInfo.username, JSON.stringify(channels));
+    var defer = q.defer();
+    clientConnection.create(server, {
+        nick: nameInfo.nick,
+        username: nameInfo.username,
+        realname: nameInfo.realname,
+        password: password
+    }, function(connInst) {
+        // === Attach bot listeners ===
+        if (clientIsBot) {
+            // make the bot listen for join/parts
+            connInst.addListener("part", function(chan, nick, reason, msg) {
+                callbacks.onPart(server, nick, chan, "part");
+            });
+            connInst.addListener("quit", function(nick, reason, chans, msg) {
+                chans = chans || [];
+                chans.forEach(function(chan) {
+                    callbacks.onPart(server, nick, chan, "quit");
+                });
+            });
+            connInst.addListener("kick", function(chan, nick, by, reason, msg) {
+                callbacks.onPart(server, nick, chan, "kick");
+            });
+            connInst.addListener("join", function(chan, nick, msg) {
+                callbacks.onJoin(server, nick, chan, "join");
+            });
+            // bucket names and drain them once per second to avoid flooding
+            // the matrix side with registrations / joins
+            var namesBucket = [
+            //  { chan: <channel>, nick: <nick> }
+            ];
+            var processingBucket = false;
+            var popName = function() {
+                var name = namesBucket.pop(); // LIFO but who cares
+                if (!name) {
+                    processingBucket = false;
+                    return;
+                }
+                log.debug(
+                    "Pop %s/%s from names bucket (%s remaining)",
+                    name.nick, name.chan, namesBucket.length
+                );
+                callbacks.onJoin(server, name.nick, name.chan, "names");
+                setTimeout(popName, 1000);
+            };
+
+            connInst.addListener("names", function(chan, names, msg) {
+                if (names) {
+                    Object.keys(names).forEach(function(nick) {
+                        namesBucket.push({
+                            chan: chan,
+                            nick: nick
+                        });
+                        // var opsLevel = names[nick]; // + @ or empty string
+                        // TODO do something with opsLevel
+                    });
+                    log.debug("Names bucket has %s entries", namesBucket.length);
+                    if (!processingBucket) {
+                        processingBucket = true;
+                        popName();
+                    }
+                }
+            });
+            // listen for mode changes
+            connInst.addListener("+mode", function(channel, by, mode, arg) {
+                callbacks.onMode(server, channel, by, mode, true, arg);
+            });
+            connInst.addListener("-mode", function(channel, by, mode, arg) {
+                callbacks.onMode(server, channel, by, mode, false, arg);
+            });
+            connInst.addListener("message", function(from, to, text) {
+                callbacks.onMessage(
+                    server, from, to, actions.irc.createMessage(text)
+                );
+            });
+            connInst.addListener("ctcp-privmsg", function(from, to, text) {
+                if (text.indexOf("ACTION ") === 0) {
+                    callbacks.onMessage(server, from, to, actions.irc.createEmote(
+                        text.substring("ACTION ".length)
+                    ));
+                }
+            });
+            connInst.addListener("notice", function(from, to, text) {
+                if (from) { // ignore server notices
+                    callbacks.onMessage(
+                        server, from, to, actions.irc.createNotice(text)
+                    );
+                }
+            });
+            connInst.addListener("topic", function(channel, topic, nick) {
+                if (nick.indexOf("@") !== -1) {
+                    var match = nick.match(
+                        // https://github.com/martynsmith/node-irc/blob/master/lib/parse_message.js#L26
+                        /^([_a-zA-Z0-9\[\]\\`^{}|-]*)(!([^@]+)@(.*))?$/
+                    );
+                    if (match) {
+                        nick = match[1];
+                    }
+                }
+                callbacks.onMessage(
+                    server, nick, channel, actions.irc.createTopic(topic)
+                );
+            });
+        }
+        // === Attach client listeners ===
+        else {
+            // just listen for PMs for clients. If you listen for rooms, you'll get
+            // duplicates since the bot will also invoke the callback fn!
+            connInst.addListener("message", function(from, to, text) {
+                if (to.indexOf("#") === 0) { return; }
+                callbacks.onMessage(
+                    server, from, to, actions.irc.createMessage(text)
+                );
+            });
+            connInst.addListener("notice", function(from, to, text) {
+                if (!from || to.indexOf("#") === 0) { return; }
+                callbacks.onMessage(
+                    server, from, to, actions.irc.createNotice(text)
+                );
+            });
+            connInst.addListener("ctcp-privmsg", function(from, to, text) {
+                if (to.indexOf("#") === 0) { return; }
+                if (text.indexOf("ACTION ") === 0) {
+                    callbacks.onMessage(
+                        server, from, to, actions.irc.createEmote(
+                            text.substring("ACTION ".length)
+                        )
+                    );
+                }
+            });
+        }
+    }).done(function(connInst) {
+        var localPort = -1;
+        if (connInst.client.conn && connInst.client.conn.localPort) {
+            localPort = connInst.client.conn.localPort;
+        }
+        if (localPort > 0) {
+            ident.setMapping(nameInfo.username, localPort);
+        }
+        // join specified channels
+        if (channels) {
+            for (var i = 0; i < channels.length; i++) {
+                connInst.client.join(channels[i]);
+            }
+        }
+        defer.resolve(connInst);
+    }, function(e) {
+        defer.reject(e);
+    });
+
+    return defer.promise;
 };
 
 VirtualIrcUser.prototype.disconnect = function(reason) {
