+++ lib/irclib/server.js
@@ -2,18 +2,8 @@
  * Represents a single IRC server.
  */
 "use strict";
-
-// The max length of <realname> in USER commands
-var MAX_REAL_NAME_LENGTH = 48;
-// The max length of <username> in USER commands
-var MAX_USER_NAME_LENGTH = 32;
-
-var q = require("q");
-var ident = require("./ident");
-var clientConnection = require("./client-connection");
 var logging = require("../logging");
 var log = logging.get("irc-server");
-var actions = require("../models/actions");
 
 /**
  * Construct a new IRC Server.
@@ -61,196 +51,6 @@ function IrcServer(domain, serviceConfig) {
         }
     }
 }
-
-/**
- * @param {Object} connectionOpts
- * @param {Object} callbacks
- * @param {Deferred=} existingDefer
- * @return {ConnectionInstance} A new connected connection instance.
- */
-IrcServer.prototype.connectAs = function(connectionOpts, callbacks, existingDefer) {
-    var channels = connectionOpts.channels;
-    var nick = connectionOpts.nick;
-    var clientIsBot = nick === this.nick;
-
-    // strip illegal chars according to RFC 1459 Sect 2.3.1
-    // but allow _ because most IRC servers allow that.
-    nick = nick.replace(/[^A-Za-z0-9\]\[\^\\\{\}\-`_]/g, "");
-    var username = connectionOpts.username || "matrixirc";
-    // real name can be any old ASCII
-    var realname = username.replace(/[^\x00-\x7F]/g, "");
-    // strip out bad characters in the username (will need to do something
-    // better like punycode with win95 style LONGNAM~1 in the future)
-    username = username.replace(/:/g, "__");
-    username = username.replace(/[^A-Za-z0-9\]\[\^\\\{\}\-`_]/g, "");
-    username = username.substring(0, MAX_USER_NAME_LENGTH);
-    realname = realname.substring(0, MAX_REAL_NAME_LENGTH);
-    var password = clientIsBot ? this.nickPass : undefined;
-
-    // strip do not track channels
-    if (channels) {
-        for (var i = 0; i < channels.length; i++) {
-            if (this.doNotTrackChannels.indexOf(channels[i]) !== -1) {
-                channels.splice(i, 1);
-                i--;
-            }
-        }
-    }
-
-    log.info("Connecting to IRC server %s as %s (user=%s)- Joining channels %s",
-        this.domain, nick, username, JSON.stringify(channels));
-    var thisServer = this;
-    var defer = q.defer();
-    clientConnection.create(this, {
-        nick: nick,
-        username: username,
-        realname: realname,
-        password: password
-    }, function(connInst) {
-        // === Attach bot listeners ===
-        if (clientIsBot) {
-            // make the bot listen for join/parts
-            connInst.addListener("part", function(chan, nick, reason, msg) {
-                callbacks.onPart(thisServer, nick, chan, "part");
-            });
-            connInst.addListener("quit", function(nick, reason, chans, msg) {
-                chans = chans || [];
-                chans.forEach(function(chan) {
-                    callbacks.onPart(thisServer, nick, chan, "quit");
-                });
-            });
-            connInst.addListener("kick", function(chan, nick, by, reason, msg) {
-                callbacks.onPart(thisServer, nick, chan, "kick");
-            });
-            connInst.addListener("join", function(chan, nick, msg) {
-                callbacks.onJoin(thisServer, nick, chan, "join");
-            });
-            // bucket names and drain them once per second to avoid flooding
-            // the matrix side with registrations / joins
-            var namesBucket = [
-            //  { chan: <channel>, nick: <nick> }
-            ];
-            var processingBucket = false;
-            var popName = function() {
-                var name = namesBucket.pop(); // LIFO but who cares
-                if (!name) {
-                    processingBucket = false;
-                    return;
-                }
-                log.debug(
-                    "Pop %s/%s from names bucket (%s remaining)",
-                    name.nick, name.chan, namesBucket.length
-                );
-                callbacks.onJoin(thisServer, name.nick, name.chan, "names");
-                setTimeout(popName, 1000);
-            };
-
-            connInst.addListener("names", function(chan, names, msg) {
-                if (names) {
-                    Object.keys(names).forEach(function(nick) {
-                        namesBucket.push({
-                            chan: chan,
-                            nick: nick
-                        });
-                        // var opsLevel = names[nick]; // + @ or empty string
-                        // TODO do something with opsLevel
-                    });
-                    log.debug("Names bucket has %s entries", namesBucket.length);
-                    if (!processingBucket) {
-                        processingBucket = true;
-                        popName();
-                    }
-                }
-            });
-            // listen for mode changes
-            connInst.addListener("+mode", function(channel, by, mode, arg) {
-                callbacks.onMode(thisServer, channel, by, mode, true, arg);
-            });
-            connInst.addListener("-mode", function(channel, by, mode, arg) {
-                callbacks.onMode(thisServer, channel, by, mode, false, arg);
-            });
-            connInst.addListener("message", function(from, to, text) {
-                callbacks.onMessage(
-                    thisServer, from, to, actions.irc.createMessage(text)
-                );
-            });
-            connInst.addListener("ctcp-privmsg", function(from, to, text) {
-                if (text.indexOf("ACTION ") === 0) {
-                    callbacks.onMessage(thisServer, from, to, actions.irc.createEmote(
-                        text.substring("ACTION ".length)
-                    ));
-                }
-            });
-            connInst.addListener("notice", function(from, to, text) {
-                if (from) { // ignore server notices
-                    callbacks.onMessage(
-                        thisServer, from, to, actions.irc.createNotice(text)
-                    );
-                }
-            });
-            connInst.addListener("topic", function(channel, topic, nick) {
-                if (nick.indexOf("@") !== -1) {
-                    var match = nick.match(
-                        // https://github.com/martynsmith/node-irc/blob/master/lib/parse_message.js#L26
-                        /^([_a-zA-Z0-9\[\]\\`^{}|-]*)(!([^@]+)@(.*))?$/
-                    );
-                    if (match) {
-                        nick = match[1];
-                    }
-                }
-                callbacks.onMessage(
-                    thisServer, nick, channel, actions.irc.createTopic(topic)
-                );
-            });
-        }
-        // === Attach client listeners ===
-        else {
-            // just listen for PMs for clients. If you listen for rooms, you'll get
-            // duplicates since the bot will also invoke the callback fn!
-            connInst.addListener("message", function(from, to, text) {
-                if (to.indexOf("#") === 0) { return; }
-                callbacks.onMessage(
-                    thisServer, from, to, actions.irc.createMessage(text)
-                );
-            });
-            connInst.addListener("notice", function(from, to, text) {
-                if (!from || to.indexOf("#") === 0) { return; }
-                callbacks.onMessage(
-                    thisServer, from, to, actions.irc.createNotice(text)
-                );
-            });
-            connInst.addListener("ctcp-privmsg", function(from, to, text) {
-                if (to.indexOf("#") === 0) { return; }
-                if (text.indexOf("ACTION ") === 0) {
-                    callbacks.onMessage(
-                        thisServer, from, to, actions.irc.createEmote(
-                            text.substring("ACTION ".length)
-                        )
-                    );
-                }
-            });
-        }
-    }).done(function(connInst) {
-        var localPort = -1;
-        if (connInst.client.conn && connInst.client.conn.localPort) {
-            localPort = connInst.client.conn.localPort;
-        }
-        if (localPort > 0) {
-            ident.setMapping(username, localPort);
-        }
-        // join specified channels
-        if (channels) {
-            for (var i = 0; i < channels.length; i++) {
-                connInst.client.join(channels[i]);
-            }
-        }
-        defer.resolve(connInst);
-    }, function(e) {
-        defer.reject(e);
-    });
-
-    return defer.promise;
-};
 
 IrcServer.prototype.hasInviteRooms = function() {
     return (
